Rename to "a11y-explicit-heading"
