+++ src/configs/recommended.js
@@ -12,7 +12,7 @@ module.exports = {
     'primer-react/no-deprecated-colors': 'warn',
     'primer-react/no-system-props': 'warn',
     'primer-react/a11y-tooltip-interactive-trigger': 'error',
-    'primer-react/explicit-heading': 'error'
+    'primer-react/a11y-explicit-heading': 'error'
   },
   settings: {
     github: {
