+++ src/rules/a11y-explicit-heading.js
